Extract broadcastable and trimmable to concerns
