+++ app/models/solid_cable/message.rb
@@ -2,19 +2,9 @@
 
 module SolidCable
   class Message < SolidCable::Record
-    scope :trimmable, lambda {
-      where(created_at: ..::SolidCable.message_retention.ago)
-    }
-    scope :broadcastable, lambda { |channels, last_id|
-      where(channel_hash: channel_hashes_for(channels)).
-        where(id: (last_id + 1)..).order(:id)
-    }
+    include Broadcastable, Trimmable
 
     class << self
-      def broadcast(channel, payload)
-        insert({ channel:, payload:, channel_hash: channel_hash_for(channel) })
-      end
-
       def channel_hashes_for(channels)
         channels.map { |channel| channel_hash_for(channel) }
       end
