+++ app/models/solid_cable/message/trimmable.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module SolidCable
+  module Message::Trimmable
+    extend ActiveSupport::Concern
+
+    included do
+      scope :trimmable, lambda {
+        where(created_at: ..::SolidCable.message_retention.ago)
+      }
+    end
+  end
+end
